ci: add examples build matrix, config variants, and upstream drift watcher
- examples job: matrix over all 5 examples (ESP-IDF v6.0.1, plus an
  early-warning release-v6.0 build for basic_connect).
- config-variants job: zerocopy, minimal, cellular-zerocopy -- the
  combinations the examples' own sdkconfig.defaults don't already
  exercise.
- Fixed examples/failover_connect/main/CMakeLists.txt, which was
  missing REQUIRES entirely.
- Fixed two real bugs this CI caught on its first real run:
  - REQUIRES cannot be conditioned on Kconfig (component-requirements
    expansion runs before Kconfig is resolved) -- the previous PR's fix
    silently no-op'd; esp_driver_uart/esp_driver_gpio are now
    unconditional REQUIRES instead.
  - A pre-existing tautological -Werror=type-limits comparison in
    ml_config_httpd.c's old-format peer migration, unrelated to this
    migration -- just never had CI to catch it before.
- Upstream drift watcher for CamM2325/microlink, same pattern as
  fugo101/wireguard-lwip's.

All 10 checks green after the fixes.

+++ components/microlink/src/ml_config_httpd.c
@@ -163,7 +163,11 @@ static void config_load_peers(ml_config_ctx_t *ctx) {
                 size_t rlen = stored_len;
                 err = nvs_get_blob(ctx->nvs, NVS_KEY_PEERS, tmp, &rlen);
                 if (err == ESP_OK) {
-                    uint8_t old_count = tmp[0];
+                    /* old_count is widened to uint16_t before the comparison: comparing the
+                     * uint8_t directly against ML_CONFIG_MAX_ALLOWED_PEERS is tautological
+                     * whenever that Kconfig value is >= 255 (true for the default of 512),
+                     * which -Werror=type-limits correctly flags as always-true. */
+                    uint16_t old_count = tmp[0];
                     uint16_t count = (old_count <= ML_CONFIG_MAX_ALLOWED_PEERS) ? old_count : 0;
                     memset(&ctx->peer_list, 0, sizeof(ctx->peer_list));
                     ctx->peer_list.count = count;
